Reorganize console index and history templates
