+++ lib/elixir_console_web/templates/console/history/index.html.leex
