+++ lib/elixir_console_web/templates/console/index.html.leex
@@ -1,0 +1,12 @@
+<div class="flex h-full flex-col sm:flex-row">
+  <div class="flex-1 sm:h-full overflow-scroll">
+    <div class="h-full flex flex-col">
+      <%= live_component(@socket, HistoryComponent, output: @output, id: :history) %>
+      <%= live_component(@socket, CommandInputComponent,
+        history: @history, bindings: @sandbox.bindings, id: :command_input) %>
+    </div>
+  </div>
+  <%= live_component(@socket, SidebarComponent,
+    sandbox: @sandbox, contextual_help: @contextual_help, suggestions: @suggestions)
+  %>
+</div>
